Moved api_check file to pkg/config; updated import paths.
